refactor(app+petcore): split overlay/AppStore hotspots, bound transport, fix localization leaks
- OverlayGeometry.swift (3.3k lines) splits into focused files:
  display-link machinery, hit-testing, interaction state, pointer
  drag, and bubble content models; the pure geometry namespace stays
- SequencedTaskSlot replaces the duplicated join/start/clear task-slot
  idiom used by runtime bootstrap and service recovery (unit tested)
- PetCoreTransport bounds in-flight requests with an async gate (16,
  below the daemon's 32-connection permit) instead of unbounded
  fan-out; waiting never blocks a dispatch thread (unit tested)
- ConnectionsModel operation results and the overlay placement
  save-failure notice resolve through the localization tables with
  full en/zh-Hans .strings + .xcstrings parity; zh copy is byte
  unchanged, en users no longer see mixed Chinese. The zh-pinned
  service-status family documents why it stays pinned
- petpack privacy tables become data-driven consts and a new test
  parses the portable skill source to reject Rust/Python drift in
  either direction
