+++ apps/macos/Sources/AgentPetCompanion/Features/Connections/ConnectionsLocalizationKeys.swift
@@ -34,6 +34,13 @@ extension APCLocalizationKey {
     static let connectionsMoreLocationsFormat = APCLocalizationKey("connections.confirm.more_locations_format", table: .connections)
     static let connectionsSafetySummary = APCLocalizationKey("connections.confirm.safety_summary", table: .connections)
     static let connectionsListTitle = APCLocalizationKey("connections.list.title", table: .connections)
+    static let connectionsCheckDoneAll = APCLocalizationKey("connections.check.done_all", table: .connections)
+    static let connectionsCheckDoneOneFormat = APCLocalizationKey("connections.check.done_one_format", table: .connections)
+    static let connectionsCheckDoneCountFormat = APCLocalizationKey("connections.check.done_count_format", table: .connections)
+    static let connectionsRepairDoneFormat = APCLocalizationKey("connections.repair.done_format", table: .connections)
+    static let connectionsRepairPendingFormat = APCLocalizationKey("connections.repair.pending_format", table: .connections)
+    static let connectionsUninstallDoneFormat = APCLocalizationKey("connections.uninstall.done_format", table: .connections)
+    static let connectionsUninstallPendingFormat = APCLocalizationKey("connections.uninstall.pending_format", table: .connections)
     static let connectionsListAccessibility = APCLocalizationKey("connections.list.accessibility", table: .connections)
     static let connectionsSourcePicker = APCLocalizationKey("connections.source_picker", table: .connections)
     static let connectionsPagePicker = APCLocalizationKey("connections.page_picker", table: .connections)
