+++ apps/macos/Sources/AgentPetCompanion/Overlay/OverlayGeometry.swift
@@ -5,912 +5,6 @@ import CoreGraphics
 import Foundation
 import QuartzCore
 
-struct OverlayDisplayGeometry: Equatable, Sendable {
-    var frame: CGRect
-    var visibleFrame: CGRect
-    var backingScaleFactor: CGFloat
-}
-
-/// The target display's presentation cadence. `CGDisplayMode.refreshRate`
-/// reflects an explicitly selected 50/60/etc. Hz mode; some built-in and
-/// adaptive displays report zero there, so AppKit's advertised maximum is the
-/// bounded fallback before the conservative 60 Hz default.
-struct OverlayDisplayRefreshCadence: Equatable, Sendable {
-    let displayID: String
-    let framesPerSecond: Double
-
-    init(displayID: String, framesPerSecond: Double) {
-        self.displayID = displayID
-        self.framesPerSecond = framesPerSecond.isFinite && framesPerSecond > 0
-            ? framesPerSecond
-            : 60
-    }
-
-    var intervalSeconds: TimeInterval {
-        1 / framesPerSecond
-    }
-
-    static func resolved(for screen: NSScreen?) -> Self {
-        let screenNumberKey = NSDeviceDescriptionKey("NSScreenNumber")
-        let screenNumber = screen?.deviceDescription[screenNumberKey] as? NSNumber
-        let displayID = screenNumber?.stringValue ?? "fallback"
-        let directDisplayID = screenNumber.map { CGDirectDisplayID($0.uint32Value) }
-        let selectedModeRefreshRate = directDisplayID
-            .flatMap(CGDisplayCopyDisplayMode)?
-            .refreshRate ?? 0
-        let advertisedMaximum = Double(screen?.maximumFramesPerSecond ?? 0)
-        let refreshRate = selectedModeRefreshRate > 0
-            ? selectedModeRefreshRate
-            : (advertisedMaximum > 0 ? advertisedMaximum : 60)
-        return Self(displayID: displayID, framesPerSecond: refreshRate)
-    }
-}
-
-/// Latest-value coalescing state shared by pointer drag and display-width
-/// preview. Callers inject monotonic timestamps, which keeps the cadence logic
-/// deterministic in tests and lets production own exactly one sleeping task.
-struct OverlayDisplayRefreshCoalescer<Value> {
-    private var pending: Value?
-    private var cadence: OverlayDisplayRefreshCadence?
-    private var lastDeliveryTime: TimeInterval?
-    private(set) var scheduledDeadline: TimeInterval?
-
-    @discardableResult
-    mutating func submit(
-        _ value: Value,
-        at time: TimeInterval,
-        cadence proposedCadence: OverlayDisplayRefreshCadence
-    ) -> TimeInterval {
-        let now = max(0, time)
-        pending = value
-        if scheduledDeadline == nil || cadence != proposedCadence {
-            let earliest: TimeInterval
-            if let lastDeliveryTime {
-                earliest = lastDeliveryTime + proposedCadence.intervalSeconds
-            } else {
-                earliest = now + proposedCadence.intervalSeconds
-            }
-            scheduledDeadline = max(now, earliest)
-        }
-        cadence = proposedCadence
-        return scheduledDeadline ?? now
-    }
-
-    mutating func takePending(deliveredAt time: TimeInterval) -> Value? {
-        guard let pending else { return nil }
-        self.pending = nil
-        scheduledDeadline = nil
-        lastDeliveryTime = max(lastDeliveryTime ?? 0, max(0, time))
-        return pending
-    }
-
-    mutating func cancelPending() {
-        pending = nil
-        scheduledDeadline = nil
-        cadence = nil
-    }
-}
-
-@MainActor
-protocol OverlayDisplayRefreshTickSource: AnyObject {
-    var isPaused: Bool { get set }
-    func invalidate()
-}
-
-@MainActor
-private final class OverlayDisplayLinkWeakTarget: NSObject {
-    private let onTick: @MainActor () -> Void
-
-    init(onTick: @escaping @MainActor () -> Void) {
-        self.onTick = onTick
-    }
-
-    @objc func displayLinkDidFire(_ displayLink: CADisplayLink) {
-        onTick()
-    }
-}
-
-@MainActor
-private final class OverlayScreenDisplayLinkTickSource:
-    OverlayDisplayRefreshTickSource
-{
-    private let target: OverlayDisplayLinkWeakTarget
-    // Source creation, mutation, and teardown are MainActor-owned. The unsafe
-    // annotation only lets deinit invalidate AppKit's non-Sendable token.
-    nonisolated(unsafe) private let displayLink: CADisplayLink
-
-    init?(screen: NSScreen?, onTick: @escaping @MainActor () -> Void) {
-        guard let screen else { return nil }
-        let target = OverlayDisplayLinkWeakTarget(onTick: onTick)
-        let displayLink = screen.displayLink(
-            target: target,
-            selector: #selector(OverlayDisplayLinkWeakTarget.displayLinkDidFire(_:))
-        )
-        self.target = target
-        self.displayLink = displayLink
-        displayLink.isPaused = true
-        displayLink.add(to: .main, forMode: .common)
-    }
-
-    var isPaused: Bool {
-        get { displayLink.isPaused }
-        set { displayLink.isPaused = newValue }
-    }
-
-    func invalidate() {
-        displayLink.invalidate()
-    }
-
-    deinit {
-        displayLink.invalidate()
-    }
-}
-
-/// Tick-driven latest-value delivery for direct manipulation. A real target
-/// screen binds an `NSScreen` display link, so adaptive-refresh panels follow
-/// actual presentation ticks rather than an assumed maximum FPS. The time
-/// coalescer remains only as the no-screen/display-link fallback.
-@MainActor
-final class OverlayDisplayLinkCoalescer<Value> {
-    typealias TickSourceFactory = @MainActor (
-        _ screen: NSScreen?,
-        _ onTick: @escaping @MainActor () -> Void
-    ) -> (any OverlayDisplayRefreshTickSource)?
-
-    private let tickSourceFactory: TickSourceFactory
-    private var tickSource: (any OverlayDisplayRefreshTickSource)?
-    private var targetDisplayID: String?
-    private var pending: Value?
-    private var delivery: ((Value) -> Void)?
-    private var fallbackCoalescer = OverlayDisplayRefreshCoalescer<Void>()
-    private var fallbackTask: Task<Void, Never>?
-    private var fallbackTaskDeadline: TimeInterval?
-    /// Keeps the tick source armed for a whole gesture. Pausing after every
-    /// delivery and re-arming on the next pointer sample can miss the upcoming
-    /// vsync, so a steady drag presents on an irregular cadence even though
-    /// every sample arrives on time.
-    private var sustainsTicks = false
-
-    init(
-        tickSourceFactory: @escaping TickSourceFactory = { screen, onTick in
-            return OverlayScreenDisplayLinkTickSource(
-                screen: screen,
-                onTick: onTick
-            )
-        }
-    ) {
-        self.tickSourceFactory = tickSourceFactory
-    }
-
-    var hasPending: Bool { pending != nil }
-
-    /// Holds the tick source unpaused until `endSustainedDelivery`, so an
-    /// active gesture presents on every display tick instead of re-arming a
-    /// paused link per pointer sample.
-    func beginSustainedDelivery() {
-        sustainsTicks = true
-        tickSource?.isPaused = false
-    }
-
-    func endSustainedDelivery() {
-        sustainsTicks = false
-        guard pending == nil else { return }
-        tickSource?.isPaused = true
-    }
-
-    func submit(
-        _ value: Value,
-        targetDisplayID: String,
-        screen: NSScreen?,
-        fallbackCadence: OverlayDisplayRefreshCadence,
-        deliver: @escaping (Value) -> Void
-    ) {
-        pending = value
-        delivery = deliver
-        if self.targetDisplayID != targetDisplayID {
-            rebind(targetDisplayID: targetDisplayID, screen: screen)
-        }
-        if let tickSource {
-            tickSource.isPaused = false
-            return
-        }
-        let deadline = fallbackCoalescer.submit(
-            (),
-            at: ProcessInfo.processInfo.systemUptime,
-            cadence: fallbackCadence
-        )
-        scheduleFallback(at: deadline)
-    }
-
-    func flushNow() {
-        deliverPending()
-    }
-
-    func cancelPending() {
-        pending = nil
-        delivery = nil
-        sustainsTicks = false
-        tickSource?.isPaused = true
-        fallbackTask?.cancel()
-        fallbackTask = nil
-        fallbackTaskDeadline = nil
-        fallbackCoalescer.cancelPending()
-    }
-
-    func invalidate() {
-        cancelPending()
-        tickSource?.invalidate()
-        tickSource = nil
-        targetDisplayID = nil
-    }
-
-    func receiveDisplayRefreshTick() {
-        deliverPending()
-    }
-
-    private func rebind(targetDisplayID: String, screen: NSScreen?) {
-        tickSource?.invalidate()
-        tickSource = nil
-        fallbackTask?.cancel()
-        fallbackTask = nil
-        fallbackTaskDeadline = nil
-        fallbackCoalescer.cancelPending()
-        self.targetDisplayID = targetDisplayID
-        tickSource = tickSourceFactory(screen) { [weak self] in
-            self?.receiveDisplayRefreshTick()
-        }
-    }
-
-    private func deliverPending() {
-        guard let pending else {
-            tickSource?.isPaused = !sustainsTicks
-            return
-        }
-        let delivery = delivery
-        self.pending = nil
-        self.delivery = nil
-        tickSource?.isPaused = !sustainsTicks
-        fallbackTask?.cancel()
-        fallbackTask = nil
-        fallbackTaskDeadline = nil
-        _ = fallbackCoalescer.takePending(
-            deliveredAt: ProcessInfo.processInfo.systemUptime
-        )
-        delivery?(pending)
-    }
-
-    private func scheduleFallback(at deadline: TimeInterval) {
-        if let fallbackTaskDeadline,
-           abs(fallbackTaskDeadline - deadline) < 0.000_001,
-           fallbackTask != nil {
-            return
-        }
-        fallbackTask?.cancel()
-        fallbackTaskDeadline = deadline
-        let delay = max(0, deadline - ProcessInfo.processInfo.systemUptime)
-        fallbackTask = Task { @MainActor [weak self] in
-            try? await Task.sleep(for: .seconds(delay))
-            guard let self, !Task.isCancelled else { return }
-            self.fallbackTask = nil
-            self.fallbackTaskDeadline = nil
-            self.deliverPending()
-        }
-    }
-}
-
-struct OverlayPetVisualEnvelope: Equatable, Sendable {
-    var canvasSize: CGSize
-    var visibleBounds: CGRect
-}
-
-/// A one-bit-per-pixel interaction mask extracted while a pet frame is
-/// decoded. Rows are stored in the same top-to-bottom order as the source
-/// `CGImage`; callers query it with bottom-left image coordinates so the
-/// conversion to the Metal renderer's coordinate system stays explicit.
-struct OverlayPetAlphaMask: Equatable, Sendable {
-    static let interactionAlphaThreshold: UInt8 = 2
-
-    let pixelWidth: Int
-    let pixelHeight: Int
-    private let opaqueBits: [UInt8]
-
-    var storageByteCount: Int { opaqueBits.count }
-
-    init?(
-        pixelWidth: Int,
-        pixelHeight: Int,
-        opaqueBits: [UInt8]
-    ) {
-        guard let requiredByteCount = Self.requiredByteCount(
-            pixelWidth: pixelWidth,
-            pixelHeight: pixelHeight
-        ), opaqueBits.count == requiredByteCount else {
-            return nil
-        }
-        self.pixelWidth = pixelWidth
-        self.pixelHeight = pixelHeight
-        self.opaqueBits = opaqueBits
-    }
-
-    /// Test/support initializer for top-to-bottom, one-byte alpha samples.
-    init?(
-        pixelWidth: Int,
-        pixelHeight: Int,
-        alphaValuesTopToBottom: [UInt8],
-        alphaThreshold: UInt8 = interactionAlphaThreshold
-    ) {
-        guard
-            let pixelCount = Self.pixelCount(pixelWidth: pixelWidth, pixelHeight: pixelHeight),
-            alphaValuesTopToBottom.count == pixelCount,
-            let byteCount = Self.requiredByteCount(
-                pixelWidth: pixelWidth,
-                pixelHeight: pixelHeight
-            )
-        else {
-            return nil
-        }
-        var bits = [UInt8](repeating: 0, count: byteCount)
-        for index in 0..<pixelCount where alphaValuesTopToBottom[index] > alphaThreshold {
-            bits[index >> 3] |= UInt8(1 << (index & 7))
-        }
-        self.init(pixelWidth: pixelWidth, pixelHeight: pixelHeight, opaqueBits: bits)
-    }
-
-    func containsOpaquePixel(atBottomLeftPoint point: CGPoint) -> Bool {
-        guard point.x.isFinite, point.y.isFinite,
-              point.x >= 0, point.y >= 0,
-              point.x < CGFloat(pixelWidth), point.y < CGFloat(pixelHeight) else {
-            return false
-        }
-        let x = Int(point.x.rounded(.down))
-        let bottomRow = Int(point.y.rounded(.down))
-        let topRow = pixelHeight - 1 - bottomRow
-        let index = topRow * pixelWidth + x
-        return opaqueBits[index >> 3] & UInt8(1 << (index & 7)) != 0
-    }
-
-    static func requiredByteCount(pixelWidth: Int, pixelHeight: Int) -> Int? {
-        guard let pixels = pixelCount(pixelWidth: pixelWidth, pixelHeight: pixelHeight) else {
-            return nil
-        }
-        let (adjusted, overflow) = pixels.addingReportingOverflow(7)
-        return overflow ? nil : adjusted / 8
-    }
-
-    private static func pixelCount(pixelWidth: Int, pixelHeight: Int) -> Int? {
-        guard pixelWidth > 0, pixelHeight > 0 else { return nil }
-        let (count, overflow) = pixelWidth.multipliedReportingOverflow(by: pixelHeight)
-        return overflow ? nil : count
-    }
-}
-
-/// Describes the exact decoded frame currently presented by the Metal view.
-/// `frameID` lets the renderer and AppStore coalesce repeated display-link
-/// draws without comparing the mask payload for every authored frame.
-struct OverlayPetFrameHitTest: Equatable, Sendable {
-    let frameID: UUID
-    let canvasSize: CGSize
-    let alphaMask: OverlayPetAlphaMask
-
-    init(
-        frameID: UUID = UUID(),
-        canvasSize: CGSize,
-        alphaMask: OverlayPetAlphaMask
-    ) {
-        self.frameID = frameID
-        self.canvasSize = canvasSize
-        self.alphaMask = alphaMask
-    }
-}
-
-enum OverlayPetAnimationIdentity {
-    static func stateEntryID(for state: ActiveAgentState?) -> String {
-        guard let state else { return "idle" }
-        if let projectedID = nonEmpty(state.overlayDisplay?.stateEntryID) {
-            return projectedID
-        }
-        let event = state.event
-        switch event.eventType {
-        case .start:
-            return "idle"
-        case .thinking, .plan:
-            let activation = nonEmpty(state.sessionActivatedAt)
-                ?? event.id
-            return scopedEntryID(
-                event: event,
-                sessionID: state.sessionID ?? event.sessionID,
-                marker: activation,
-                reaction: "thinking"
-            )
-        case .done:
-            let completion = nonEmpty(state.sessionActivatedAt)
-                ?? event.id
-            return scopedEntryID(
-                event: event,
-                sessionID: state.sessionID ?? event.sessionID,
-                marker: completion,
-                reaction: "done"
-            )
-        case .tool, .waiting, .failed:
-            // PetCore owns the projected identity, including the tool activity
-            // run that lets genuinely new tool work replay its entry burst. A
-            // single event carries no run history, so this compatibility
-            // fallback keeps one stable identity per reaction.
-            return event.eventType.rawValue
-        }
-    }
-
-    private static func scopedEntryID(
-        event: AgentEvent,
-        sessionID: String?,
-        marker: String,
-        reaction: String
-    ) -> String {
-        [
-            reaction,
-            event.source.rawValue,
-            nonEmpty(sessionID),
-            marker
-        ]
-        .compactMap { $0 }
-        .joined(separator: ":")
-    }
-
-    private static func nonEmpty(_ value: String?) -> String? {
-        let value = value?.trimmingCharacters(in: .whitespacesAndNewlines)
-        return value?.isEmpty == false ? value : nil
-    }
-
-}
-
-enum OverlayControlVisibility {
-    static let hoverShowDelay = Duration.zero
-    static let hoverHideDelay = Duration.milliseconds(300)
-
-    static func isVisible(
-        pointerNearPet: Bool,
-        petDragInProgress: Bool,
-        keyboardFocusActive: Bool = false
-    ) -> Bool {
-        pointerNearPet || petDragInProgress || keyboardFocusActive
-    }
-
-    static func transitionDelay(showing: Bool, forced: Bool) -> Duration {
-        forced ? .zero : (showing ? hoverShowDelay : hoverHideDelay)
-    }
-}
-
-/// Motion values shared by SwiftUI content and the AppKit overlay panels.
-enum OverlayMotion {
-    static let controlFadeDuration: TimeInterval = 0.14
-    static let controlFadeDelay = Duration.milliseconds(140)
-    static let bubbleLayoutDuration: TimeInterval = 0.16
-    static let bubbleLayoutDelay = Duration.milliseconds(160)
-    static let reducedMotionCrossfadeDuration: TimeInterval = 0.16
-    static let reducedMotionCrossfadeDelay = Duration.milliseconds(160)
-    static let reducedMotionCrossfadeHalfDelay = Duration.milliseconds(80)
-}
-
-/// One presentation state coordinates every transient overlay control. Pointer
-/// hover begins the visual fade immediately, while pointer exit is delayed to
-/// avoid flicker between the pet and its controls. Keyboard focus and an active
-/// pet drag remain immediate.
-@MainActor
-final class OverlayControlPresentationState: ObservableObject {
-    typealias TransitionSleeper = @MainActor @Sendable (Duration) async throws -> Void
-
-    enum Region: Hashable {
-        case pet
-        case bubble
-        case menu
-    }
-
-    @Published private(set) var isVisible = false
-    @Published private(set) var keyboardNavigationActive = false
-    var visibilityDidChange: (() -> Void)?
-
-    private var hoveredRegions: Set<Region> = []
-    private var focusedRegions: Set<Region> = []
-    private var activeRegions: Set<Region> = []
-    private var transitionTask: Task<Void, Never>?
-    private let transitionSleeper: TransitionSleeper
-
-    init(
-        transitionSleeper: @escaping TransitionSleeper = { delay in
-            try await Task.sleep(for: delay)
-        }
-    ) {
-        self.transitionSleeper = transitionSleeper
-    }
-
-    func setHovered(_ region: Region, _ hovered: Bool) {
-        update(region, enabled: hovered, in: &hoveredRegions)
-        scheduleVisibilityUpdate()
-    }
-
-    func setFocused(_ region: Region, _ focused: Bool) {
-        update(region, enabled: focused, in: &focusedRegions)
-        let nextKeyboardNavigationActive = !focusedRegions.isEmpty
-        if keyboardNavigationActive != nextKeyboardNavigationActive {
-            keyboardNavigationActive = nextKeyboardNavigationActive
-        }
-        scheduleVisibilityUpdate()
-    }
-
-    func setActive(_ region: Region, _ active: Bool) {
-        update(region, enabled: active, in: &activeRegions)
-        scheduleVisibilityUpdate()
-    }
-
-    func reset() {
-        transitionTask?.cancel()
-        transitionTask = nil
-        hoveredRegions.removeAll()
-        focusedRegions.removeAll()
-        activeRegions.removeAll()
-        keyboardNavigationActive = false
-        setVisible(false)
-    }
-
-    private func update(
-        _ region: Region,
-        enabled: Bool,
-        in regions: inout Set<Region>
-    ) {
-        if enabled {
-            regions.insert(region)
-        } else {
-            regions.remove(region)
-        }
-    }
-
-    private func scheduleVisibilityUpdate() {
-        transitionTask?.cancel()
-        let shouldShow = !hoveredRegions.isEmpty
-            || !focusedRegions.isEmpty
-            || !activeRegions.isEmpty
-        let forced = !focusedRegions.isEmpty || !activeRegions.isEmpty
-        guard shouldShow != isVisible else { return }
-        let delay = OverlayControlVisibility.transitionDelay(
-            showing: shouldShow,
-            forced: forced
-        )
-        if delay == .zero {
-            setVisible(shouldShow)
-            return
-        }
-        let transitionSleeper = transitionSleeper
-        transitionTask = Task { @MainActor [weak self] in
-            do {
-                try await transitionSleeper(delay)
-            } catch {
-                return
-            }
-            guard let self, !Task.isCancelled else { return }
-            let latestShouldShow = !self.hoveredRegions.isEmpty
-                || !self.focusedRegions.isEmpty
-                || !self.activeRegions.isEmpty
-            guard latestShouldShow == shouldShow else { return }
-            self.setVisible(shouldShow)
-        }
-    }
-
-    private func setVisible(_ visible: Bool) {
-        guard isVisible != visible else { return }
-        isVisible = visible
-        visibilityDidChange?()
-    }
-}
-
-/// Keeps high-frequency display-size preview out of `AppStore`. Only the pet
-/// canvas observes this state, so slider movement does not invalidate the
-/// control center, bubbles, or unrelated overlay content.
-@MainActor
-final class OverlayInteractionPresentationState: ObservableObject {
-    private struct DisplayWidthPresentation: Equatable {
-        var displayWidthPt: CGFloat
-        var petLocalCenter: CGPoint
-    }
-
-    @Published private var displayWidthPresentation: DisplayWidthPresentation?
-    @Published private(set) var petInteraction: OverlayPetInteractionPresentation?
-    @Published private(set) var pressFeedbackActive = false
-    private var dragInteractionID: UUID?
-    private var lastDragCenter: CGPoint?
-
-    func resolvedDisplayWidthPt(fallback: CGFloat) -> CGFloat {
-        displayWidthPresentation?.displayWidthPt ?? fallback
-    }
-
-    func resolvedPetLocalCenter(fallback: CGPoint) -> CGPoint {
-        displayWidthPresentation?.petLocalCenter ?? fallback
-    }
-
-    func present(displayWidthPt: CGFloat, petLocalCenter: CGPoint) {
-        let next = DisplayWidthPresentation(
-            displayWidthPt: displayWidthPt,
-            petLocalCenter: petLocalCenter
-        )
-        guard displayWidthPresentation != next else { return }
-        displayWidthPresentation = next
-    }
-
-    func clearDisplayWidthPreview() {
-        guard displayWidthPresentation != nil else { return }
-        displayWidthPresentation = nil
-    }
-
-    func acknowledge(reduceMotion: Bool) {
-        guard !reduceMotion, dragInteractionID == nil else { return }
-        let interactionID = UUID()
-        petInteraction = OverlayPetInteractionPresentation(
-            stateName: "acknowledge",
-            entryID: "interaction:\(interactionID.uuidString):acknowledge"
-        )
-    }
-
-    func beginPressFeedback(enabled: Bool, reduceMotion: Bool) {
-        guard enabled, !reduceMotion, dragInteractionID == nil else { return }
-        pressFeedbackActive = true
-    }
-
-    func beginDrag(interactionID: UUID, center: CGPoint) {
-        dragInteractionID = interactionID
-        lastDragCenter = center
-        petInteraction = nil
-    }
-
-    func updateDrag(interactionID: UUID, center: CGPoint) {
-        guard dragInteractionID == interactionID else { return }
-        pressFeedbackActive = false
-        let priorCenter = lastDragCenter ?? center
-        lastDragCenter = center
-        let horizontalDelta = center.x - priorCenter.x
-        guard abs(horizontalDelta) >= OverlayPetInteractionPolicy.directionThresholdPt else {
-            return
-        }
-        let stateName = horizontalDelta < 0 ? "drag_left" : "drag_right"
-        guard petInteraction?.stateName != stateName else { return }
-        petInteraction = OverlayPetInteractionPresentation(
-            stateName: stateName,
-            entryID: "interaction:\(interactionID.uuidString):\(stateName)"
-        )
-    }
-
-    func endDrag(interactionID: UUID?) {
-        if let interactionID, dragInteractionID != interactionID { return }
-        dragInteractionID = nil
-        lastDragCenter = nil
-        pressFeedbackActive = false
-        if petInteraction?.stateName == "drag_left"
-            || petInteraction?.stateName == "drag_right"
-        {
-            petInteraction = nil
-        }
-    }
-
-    func complete(entryID: String) {
-        guard petInteraction?.entryID == entryID,
-              petInteraction?.stateName == "acknowledge"
-        else { return }
-        petInteraction = nil
-    }
-}
-
-struct OverlayPetInteractionPresentation: Equatable, Sendable {
-    var stateName: String
-    var entryID: String
-}
-
-enum OverlayPetInteractionPolicy {
-    static let directionThresholdPt: CGFloat = 0.5
-    static let pressFeedbackDurationSeconds = 0.16
-}
-
-enum OverlayPointerMaskState: Equatable, Sendable {
-    case missing
-    case valid
-    case stale
-}
-
-struct OverlayPointerOwnershipInput: Equatable, Sendable {
-    var overlayVisible: Bool
-    var activeInteractionID: UUID?
-    var maskState: OverlayPointerMaskState
-    var validMaskPixelIsOpaque: Bool
-    var pointerInBubble: Bool
-    var pointerInMenu: Bool
-    var pointerInGeometricPetRegion: Bool
-}
-
-enum OverlayPointerOwnership: Equatable, Sendable {
-    case passthrough
-    case pet
-    case auxiliarySurface
-    case activeLease(UUID)
-
-    var isOwnedByOverlay: Bool {
-        self != .passthrough
-    }
-}
-
-/// The only policy that decides whether the overlay owns a pointer location.
-/// AppKit event routing remains a defensive delivery mechanism; it does not
-/// redefine mask fallback or active-gesture ownership.
-enum OverlayPointerOwnershipPolicy {
-    static func resolve(
-        _ input: OverlayPointerOwnershipInput
-    ) -> OverlayPointerOwnership {
-        guard input.overlayVisible else { return .passthrough }
-        if let interactionID = input.activeInteractionID {
-            return .activeLease(interactionID)
-        }
-        if input.pointerInBubble || input.pointerInMenu {
-            return .auxiliarySurface
-        }
-        guard input.pointerInGeometricPetRegion else {
-            return .passthrough
-        }
-        switch input.maskState {
-        case .valid:
-            return input.validMaskPixelIsOpaque ? .pet : .passthrough
-        case .missing, .stale:
-            return .pet
-        }
-    }
-}
-
-enum OverlayPetPointerGesture {
-    static let dragThreshold: CGFloat = 4
-
-    static func exceedsDragThreshold(from start: CGPoint, to current: CGPoint) -> Bool {
-        hypot(current.x - start.x, current.y - start.y) >= dragThreshold
-    }
-
-    /// AppKit reports a double-click as click counts one and two. Performing
-    /// the primary action only for the first release preserves responsive
-    /// single-click behavior without toggling the bubble back on the second.
-    static func shouldPerformPrimaryClick(clickCount: Int, didDrag: Bool) -> Bool {
-        !didDrag && clickCount == 1
-    }
-}
-
-/// Converts a pointer sample out of CoreGraphics' global event space into the
-/// AppKit screen space the drag anchor uses. A drag translates the host window
-/// on every display tick, so a sample derived from the window's own coordinate
-/// system silently subtracts a move it never observed; taking the event's
-/// absolute location instead keeps the grab point fixed under the pointer no
-/// matter how far or how fast the composition has moved.
-enum OverlayPointerCoordinateSpace {
-    static func screenPoint(
-        forGlobalEventLocation location: CGPoint,
-        zeroOriginScreenFrame: CGRect
-    ) -> CGPoint {
-        CGPoint(
-            x: location.x,
-            y: zeroOriginScreenFrame.maxY - location.y
-        )
-    }
-}
-
-enum OverlayPointerGesturePhase: Equatable, Sendable {
-    case idle
-    case pressed
-    case dragging
-    case finalized
-}
-
-struct OverlayDragSession: Equatable, Sendable {
-    let interactionID: UUID
-    let startPointerScreen: CGPoint
-    let startAnchorScreen: CGPoint
-    var latestPointerScreen: CGPoint
-    let startDisplayID: String
-    var hasCrossedThreshold: Bool
-    private(set) var phase: OverlayPointerGesturePhase
-
-    init(
-        interactionID: UUID = UUID(),
-        startPointerScreen: CGPoint,
-        startAnchorScreen: CGPoint,
-        startDisplayID: String
-    ) {
-        self.interactionID = interactionID
-        self.startPointerScreen = startPointerScreen
-        self.startAnchorScreen = startAnchorScreen
-        latestPointerScreen = startPointerScreen
-        self.startDisplayID = startDisplayID
-        hasCrossedThreshold = false
-        phase = .pressed
-    }
-
-    var proposedAnchorScreen: CGPoint {
-        CGPoint(
-            x: startAnchorScreen.x
-                + latestPointerScreen.x
-                - startPointerScreen.x,
-            y: startAnchorScreen.y
-                + latestPointerScreen.y
-                - startPointerScreen.y
-        )
-    }
-
-    mutating func updatePointer(_ point: CGPoint) {
-        guard phase != .finalized else { return }
-        latestPointerScreen = point
-        if !hasCrossedThreshold {
-            hasCrossedThreshold = OverlayPetPointerGesture.exceedsDragThreshold(
-                from: startPointerScreen,
-                to: point
-            )
-            if hasCrossedThreshold {
-                phase = .dragging
-            }
-        }
-    }
-
-    mutating func compareAndFinalize(interactionID: UUID) -> Bool {
-        guard self.interactionID == interactionID,
-              phase != .finalized else { return false }
-        phase = .finalized
-        return true
-    }
-}
-
-struct OverlayDragDisplay: Equatable, Sendable {
-    let id: String
-    let frame: CGRect
-    let visibleFrame: CGRect
-}
-
-/// Display geometry and presentation cadence captured once per gesture rather
-/// than per pointer sample. Rebuilding the screen list and asking CoreGraphics
-/// for the selected display mode on every event costs a window-server round
-/// trip inside the pointer handler, which is exactly the per-event main-thread
-/// work that makes a fast drag stutter. The desktop arrangement only changes
-/// through a screen-parameter notification, so a gesture-scoped snapshot stays
-/// authoritative for the whole drag.
-struct OverlayDragDisplayCatalog: Equatable, Sendable {
-    let displays: [OverlayDragDisplay]
-    let fallbackCadence: OverlayDisplayRefreshCadence
-    private let cadences: [String: OverlayDisplayRefreshCadence]
-
-    init(
-        displays: [OverlayDragDisplay],
-        cadences: [String: OverlayDisplayRefreshCadence],
-        fallbackCadence: OverlayDisplayRefreshCadence
-    ) {
-        self.displays = displays
-        self.cadences = cadences
-        self.fallbackCadence = fallbackCadence
-    }
-
-    func cadence(for displayID: String?) -> OverlayDisplayRefreshCadence {
-        guard let displayID, let cadence = cadences[displayID] else {
-            return fallbackCadence
-        }
-        return cadence
-    }
-}
-
-/// Selects the drag display from injected geometry so crossing a display edge
-/// does not replace the absolute pointer anchor with window-relative deltas.
-enum OverlayDragScreenResolver {
-    static func resolve(
-        pointer: CGPoint,
-        proposedPetCenter: CGPoint,
-        displays: [OverlayDragDisplay],
-        fallbackDisplayID: String?
-    ) -> OverlayDragDisplay? {
-        displays.first { $0.frame.contains(pointer) }
-            ?? displays.first { $0.frame.contains(proposedPetCenter) }
-            ?? fallbackDisplayID.flatMap { fallbackID in
-                displays.first { $0.id == fallbackID }
-            }
-            ?? displays.first
-    }
-}
-
 enum OverlayInteractionWindowRole: Equatable, Sendable {
     case parentPetPanel
     case bubbleChildPanel
@@ -2469,816 +1563,5 @@ enum OverlayGeometry {
             max(canonicalValue, canonicalLower),
             canonicalUpper
         ))
-    }
-}
-
-enum OverlaySessionGroupTone: Int, CaseIterable, Equatable {
-    case running = 0
-    case ready = 1
-    case failed = 2
-    case needsInput = 3
-
-    init(eventType: AgentEventKind?) {
-        self = switch eventType {
-        case .waiting: .needsInput
-        case .failed: .failed
-        case .done: .ready
-        case .start, .thinking, .plan, .tool, nil: .running
-        }
-    }
-
-    static func aggregate(_ sessions: [OverlaySessionContent]) -> OverlaySessionGroupTone {
-        sessions
-            .map { OverlaySessionGroupTone(eventType: $0.eventType) }
-            .max(by: { $0.rawValue < $1.rawValue })
-            ?? .running
-    }
-}
-
-enum OverlayBubbleProjection {
-    static func contents(
-        states: [ActiveAgentState],
-        omittedCount: Int,
-        dismissedSessionIDs: Set<String>,
-        groupSessionsByAgent: Bool = false,
-        standaloneSessionOrder: [String] = [],
-        standaloneStackExpanded: Bool = true,
-        isExpanded: (AgentSource) -> Bool
-    ) -> [OverlayBubbleContent] {
-        let visibleStates = states.filter {
-            !dismissedSessionIDs.contains(OverlaySessionContent.stableID(
-                source: $0.source,
-                sessionID: $0.sessionID ?? $0.event.sessionID,
-                anonymousSessionAlias: $0.anonymousSessionAlias,
-                fallbackEventID: $0.event.id
-            ))
-        }
-        if !groupSessionsByAgent {
-            let stateByID = Dictionary(
-                visibleStates.map { state in
-                    (
-                        OverlaySessionContent.stableID(
-                            source: state.source,
-                            sessionID: state.sessionID ?? state.event.sessionID,
-                            anonymousSessionAlias: state.anonymousSessionAlias,
-                            fallbackEventID: state.event.id
-                        ),
-                        state
-                    )
-                },
-                uniquingKeysWith: { _, latest in latest }
-            )
-            var orderedStates = standaloneSessionOrder.compactMap { stateByID[$0] }
-            let orderedIDs = Set(orderedStates.map {
-                OverlaySessionContent.stableID(
-                    source: $0.source,
-                    sessionID: $0.sessionID ?? $0.event.sessionID,
-                    anonymousSessionAlias: $0.anonymousSessionAlias,
-                    fallbackEventID: $0.event.id
-                )
-            })
-            // Direct projection callers may not own App presentation state.
-            // Use PetCore's canonical attention/latest-first order as the
-            // fallback so the first expanded card matches the folded card.
-            orderedStates.append(contentsOf: visibleStates.filter {
-                !orderedIDs.contains(OverlaySessionContent.stableID(
-                    source: $0.source,
-                    sessionID: $0.sessionID ?? $0.event.sessionID,
-                    anonymousSessionAlias: $0.anonymousSessionAlias,
-                    fallbackEventID: $0.event.id
-                ))
-            })
-            // The visual list is priority-first. Stable presentation order is
-            // retained inside a tone so ordinary thinking/tool churn cannot
-            // make concurrent cards hop.
-            orderedStates = orderedStates.enumerated().sorted { left, right in
-                let leftTone = OverlaySessionGroupTone(eventType: left.element.event.eventType)
-                let rightTone = OverlaySessionGroupTone(eventType: right.element.event.eventType)
-                if leftTone != rightTone {
-                    return leftTone.rawValue > rightTone.rawValue
-                }
-                return left.offset < right.offset
-            }.map(\.element)
-
-            let representedSessionCount = orderedStates.count + omittedCount
-            if !standaloneStackExpanded,
-               representedSessionCount > 1,
-               let primaryState = orderedStates.first
-            {
-                return [OverlayBubbleContent(
-                    standaloneState: primaryState,
-                    stackSessionCount: representedSessionCount,
-                    isStackExpanded: false
-                )]
-            }
-            var cards = orderedStates.enumerated().map { index, state in
-                OverlayBubbleContent(
-                    standaloneState: state,
-                    stackSessionCount: index == 0
-                        ? representedSessionCount
-                        : 1,
-                    isStackExpanded: true
-                )
-            }
-            if omittedCount > 0 {
-                cards.append(.omittedSummary(count: omittedCount))
-            }
-            return cards
-        }
-
-        var grouped = AgentSource.allCases.compactMap { source -> OverlayBubbleContent? in
-            let sourceStates = visibleStates.filter { $0.source == source }
-            return sourceStates.isEmpty ? nil : OverlayBubbleContent(
-                source: source,
-                states: sourceStates,
-                isExpanded: isExpanded(source)
-            )
-        }
-        if omittedCount > 0 {
-            grouped.append(.omittedSummary(count: omittedCount))
-        }
-        // The pet itself communicates idle. No session means no bubble.
-        return grouped
-    }
-}
-
-enum OverlaySessionSurfaceKind: Equatable {
-    case app
-    case cli
-}
-
-enum OverlaySessionNavigationNotice: Equatable, Sendable {
-    case unavailable
-    case failed
-    case degradedToHost
-
-    func localizedText(
-        locale: String = APCLocalization.interfaceLocaleIdentifier
-    ) -> String {
-        switch self {
-        case .unavailable:
-            APCLocalization.text(.overlaySessionNavigationUnavailable, locale: locale)
-        case .failed:
-            APCLocalization.text(.overlaySessionNavigationFailed, locale: locale)
-        case .degradedToHost:
-            APCLocalization.text(.overlaySessionNavigationDegraded, locale: locale)
-        }
-    }
-}
-
-struct OverlaySessionContent: Equatable, Identifiable {
-    var id: String
-    var eventID: String
-    var source: AgentSource?
-    var sessionID: String?
-    var eventType: AgentEventKind?
-    var sessionTitle: String
-    var activityText: String
-    var messageText: String
-    var statusText: String
-    var acknowledgementID: String?
-    var navigation: AgentSessionNavigation
-    var navigationNotice: OverlaySessionNavigationNotice?
-
-    var needsUserAttention: Bool {
-        eventType == .waiting || eventType == .failed
-    }
-
-    var navigationCapability: NavigationCapability {
-        AgentSessionRouter.validatedCapability(
-            source: source,
-            sessionID: sessionID,
-            navigation: navigation
-        )
-    }
-    var canOpen: Bool {
-        source == nil || navigationCapability != .unavailable
-    }
-    var actionLabel: String {
-        actionLabel()
-    }
-    func actionLabel(
-        locale: String = APCLocalization.interfaceLocaleIdentifier
-    ) -> String {
-        guard let source else {
-            return APCLocalization.text(.overlayActionOpen, locale: locale)
-        }
-        return APCLocalizedPresentation.navigationActionTitle(
-            navigationCapability,
-            source: source,
-            navigation: navigation,
-            locale: locale
-        ) ?? APCLocalizedPresentation.navigationUnavailableTitle(locale: locale)
-    }
-    var surfaceKind: OverlaySessionSurfaceKind? {
-        switch navigation.surface {
-        case "chatgpt_app", "claude_app", "opencode_app":
-            .app
-        case "cli_terminal":
-            .cli
-        default:
-            nil
-        }
-    }
-    var surfaceLabel: String? {
-        surfaceLabel()
-    }
-    func surfaceLabel(
-        locale: String = APCLocalization.interfaceLocaleIdentifier
-    ) -> String? {
-        surfaceKind.map {
-            APCLocalizedPresentation.sessionSurfaceTitle($0, locale: locale)
-        }
-    }
-    var primaryDetailText: String {
-        detailCandidates.first ?? ""
-    }
-    var standaloneSummaryText: String {
-        primaryDetailText
-    }
-    var detailText: String {
-        primaryDetailText
-    }
-    var accessibilityReadingOrder: [String] {
-        [
-            source?.title,
-            surfaceLabel,
-            sessionTitle,
-            statusText,
-            primaryDetailText,
-            accessibilityFallbackDetailText,
-            actionLabel,
-        ]
-        .compactMap(Self.compactMessage)
-    }
-    var accessibilityLabel: String {
-        accessibilityReadingOrder.joined(separator: ", ")
-    }
-    private var detailCandidates: [String] {
-        if let navigationNotice {
-            return [navigationNotice.localizedText()]
-        }
-        if let message = Self.compactMessage(messageText) {
-            return [message]
-        }
-        if let activity = Self.compactMessage(activityText) {
-            return [activity]
-        }
-        return []
-    }
-    /// Preserves a complete VoiceOver state-to-action reading order when
-    /// PetCore intentionally omits displayable session copy. This text is
-    /// localized product guidance, never raw connector event detail, and is
-    /// not inserted into the visual bubble.
-    private var accessibilityFallbackDetailText: String? {
-        guard detailCandidates.isEmpty else { return nil }
-        let key: APCLocalizationKey? = switch eventType {
-        case .start:
-            .overlayDetailRunning
-        case .thinking:
-            .overlayActivityThinking
-        case .plan:
-            .overlayActivityPlan
-        case .tool:
-            .overlayActivityTool
-        case .waiting:
-            .overlayDetailNeedsInput
-        case .done:
-            .overlayDetailCompleted
-        case .failed:
-            .overlayDetailBlocked
-        case nil:
-            nil
-        }
-        guard let key else { return nil }
-        return Self.compactMessage(Self.nonredundantDetail(
-            APCLocalization.text(key),
-            title: sessionTitle,
-            status: statusText
-        ))
-    }
-    var dismissesAfterActivation: Bool {
-        switch eventType {
-        case .done: true
-        case .start, .thinking, .plan, .tool, .waiting, .failed, nil: false
-        }
-    }
-
-    /// Geometry-only fixture for callers that do not have live bubble
-    /// content. AppStore never publishes this as a user-visible session.
-    static let measurementPlaceholder = OverlaySessionContent(
-        id: "measurement-placeholder",
-        eventID: "measurement-placeholder",
-        source: nil,
-        sessionID: nil,
-        eventType: nil,
-        sessionTitle: "Agent Pet Companion",
-        activityText: "",
-        messageText: "",
-        statusText: "",
-        navigation: AgentSessionNavigation()
-    )
-
-    static func omittedSummary(count: Int) -> OverlaySessionContent {
-        OverlaySessionContent(
-            id: "omitted-session-summary",
-            eventID: "omitted-session-summary",
-            source: nil,
-            sessionID: nil,
-            eventType: nil,
-            sessionTitle: APCLocalization.text(.overlayMoreSessionsTitle),
-            activityText: "",
-            messageText: APCLocalization.format(.overlayMoreSessionsDetailFormat, count),
-            statusText: "",
-            navigation: AgentSessionNavigation()
-        )
-    }
-
-    init(
-        id: String,
-        eventID: String? = nil,
-        source: AgentSource?,
-        sessionID: String?,
-        eventType: AgentEventKind?,
-        sessionTitle: String,
-        activityText: String? = nil,
-        messageText: String,
-        statusText: String,
-        acknowledgementID: String? = nil,
-        navigation: AgentSessionNavigation = AgentSessionNavigation(),
-        navigationNotice: OverlaySessionNavigationNotice? = nil
-    ) {
-        self.id = id
-        self.eventID = eventID ?? id
-        self.source = source
-        self.sessionID = sessionID
-        self.eventType = eventType
-        self.sessionTitle = sessionTitle
-        self.activityText = activityText ?? ""
-        self.messageText = messageText
-        self.statusText = statusText
-        self.acknowledgementID = acknowledgementID
-        self.navigation = navigation
-        self.navigationNotice = navigationNotice
-    }
-
-    init(state: ActiveAgentState) {
-        let event = state.event
-        let resolvedSessionID = state.sessionID ?? event.sessionID
-        id = Self.stableID(
-            source: event.source,
-            sessionID: resolvedSessionID,
-            anonymousSessionAlias: state.anonymousSessionAlias,
-            fallbackEventID: event.id
-        )
-        eventID = event.id
-        source = event.source
-        sessionID = resolvedSessionID
-        eventType = event.eventType
-        acknowledgementID = state.acknowledgementID
-        statusText = Self.displayStatus(
-            for: state.overlayDisplay?.summaryKind
-                ?? Self.summaryKind(for: event.eventType)
-        )
-        let proposedTitle = Self.sessionTitle(for: state)
-        sessionTitle = Self.normalizedText(proposedTitle) == Self.normalizedText(statusText)
-            ? Self.genericSessionTitle(for: state)
-            : proposedTitle
-        navigation = state.overlayDisplay?.navigation ?? AgentSessionNavigation()
-        navigationNotice = nil
-        activityText = Self.nonredundantDetail(
-            Self.activityDetail(for: state),
-            title: sessionTitle,
-            status: statusText
-        )
-        messageText = Self.nonredundantMessage(
-            Self.assistantMessage(for: state) ?? "",
-            title: sessionTitle,
-            status: statusText
-        )
-    }
-
-    init(event: AgentEvent) {
-        id = Self.stableID(
-            source: event.source,
-            sessionID: event.sessionID,
-            fallbackEventID: event.id
-        )
-        eventID = event.id
-        source = event.source
-        sessionID = event.sessionID
-        eventType = event.eventType
-        acknowledgementID = nil
-        sessionTitle = APCLocalization.format(.overlaySessionTitleFormat, event.source.shortTitle)
-        statusText = Self.displayStatus(for: Self.summaryKind(for: event.eventType))
-        navigation = event.sessionNavigation
-        navigationNotice = nil
-        activityText = ""
-        messageText = ""
-    }
-
-    static func stableID(
-        source: AgentSource,
-        sessionID: String?,
-        anonymousSessionAlias: String? = nil,
-        fallbackEventID _: String
-    ) -> String {
-        let sessionID = sessionID?.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard let sessionID, !sessionID.isEmpty else {
-            if let anonymousSessionAlias = validatedAnonymousAlias(anonymousSessionAlias) {
-                return "session-\(source.rawValue)-\(anonymousSessionAlias)"
-            }
-            // PetCore groups unattributed events into one source-scoped session.
-            // Match that identity here so hook revisions update the existing row
-            // instead of removing/reinserting it and clearing manual dismissal.
-            return "session-\(source.rawValue)-unattributed"
-        }
-        return "session-\(source.rawValue)-\(sessionID)"
-    }
-
-    static func reopenID(for state: ActiveAgentState) -> String {
-        let stableID = stableID(
-            source: state.source,
-            sessionID: state.sessionID ?? state.event.sessionID,
-            anonymousSessionAlias: state.anonymousSessionAlias,
-            fallbackEventID: state.event.id
-        )
-        switch state.event.eventType {
-        case .waiting, .failed:
-            return "attention:\(stableID):\(state.event.id)"
-        case .start, .thinking, .plan, .tool, .done:
-            return "activation:\(stableID):\(state.sessionActivatedAt ?? "initial")"
-        }
-    }
-
-    private static func sessionTitle(for state: ActiveAgentState) -> String {
-        if let title = compactTitle(state.sessionTitle) {
-            return title
-        }
-        if state.sessionUserMessage?.role == "user",
-           let title = compactTitle(state.sessionUserMessage?.content)
-        {
-            return title
-        }
-        return genericSessionTitle(for: state)
-    }
-
-    private static func genericSessionTitle(for state: ActiveAgentState) -> String {
-        if let label = anonymousAliasLabel(state.anonymousSessionAlias) {
-            return APCLocalization.format(
-                .overlaySessionAliasTitleFormat,
-                state.source.shortTitle,
-                label
-            )
-        }
-        return APCLocalization.format(.overlaySessionTitleFormat, state.source.shortTitle)
-    }
-
-    static func anonymousAliasLabel(_ value: String?) -> String? {
-        guard let value = validatedAnonymousAlias(value),
-              value.hasPrefix("anon-"),
-              let sequence = UInt64(value.dropFirst(5), radix: 36),
-              sequence > 0
-        else {
-            return nil
-        }
-
-        var index = sequence
-        var characters: [Character] = []
-        while index > 0 {
-            index -= 1
-            let scalar = UnicodeScalar(65 + Int(index % 26))!
-            characters.append(Character(scalar))
-            index /= 26
-        }
-        return String(characters.reversed())
-    }
-
-    private static func validatedAnonymousAlias(_ value: String?) -> String? {
-        guard let value = value?.trimmingCharacters(in: .whitespacesAndNewlines),
-              value.range(
-                  of: "^anon-[0-9a-z]{1,13}$",
-                  options: .regularExpression
-              ) != nil
-        else {
-            return nil
-        }
-        return value
-    }
-
-    private static func activityDetail(for state: ActiveAgentState) -> String? {
-        // Only PetCore's bounded, normalized projection is displayable. Raw
-        // connector payloads may contain commands, paths, or serialized JSON.
-        compactMessage(state.sessionActivity?.content)
-    }
-
-    private static func assistantMessage(for state: ActiveAgentState) -> String? {
-        guard state.sessionMessage?.role == "assistant" else { return nil }
-        return compactMessage(state.sessionMessage?.content)
-    }
-
-    private static func compactTitle(_ value: String?) -> String? {
-        guard let value = compactMessage(value) else { return nil }
-        let firstLine = value.split(whereSeparator: \.isNewline).first.map(String.init) ?? value
-        return firstLine.count > 80 ? "\(firstLine.prefix(79))…" : firstLine
-    }
-
-    private static func compactMessage(_ value: String?) -> String? {
-        let value = value?.trimmingCharacters(in: .whitespacesAndNewlines)
-        return value.flatMap { $0.isEmpty ? nil : $0 }
-    }
-
-    /// Compiled once instead of per call. `replacingOccurrences(options:
-    /// .regularExpression)` builds a fresh `NSRegularExpression` every time,
-    /// and this normalization runs several times per session on every bubble
-    /// projection. The pattern and match options are unchanged, so the
-    /// resulting text is identical.
-    private static let redundancyNormalizationPattern = try? NSRegularExpression(
-        pattern: "[\\s\\p{P}]+"
-    )
-
-    private static func normalizedText(_ value: String) -> String {
-        let folded = value
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-            .folding(options: [.caseInsensitive, .diacriticInsensitive], locale: .current)
-        guard let pattern = redundancyNormalizationPattern else { return folded }
-        return pattern.stringByReplacingMatches(
-            in: folded,
-            range: NSRange(folded.startIndex..., in: folded),
-            withTemplate: ""
-        )
-    }
-
-    private static func nonredundantMessage(
-        _ proposed: String,
-        title: String,
-        status: String
-    ) -> String {
-        let occupied = Set([title, status].map(normalizedText))
-        return occupied.contains(normalizedText(proposed)) ? "" : proposed
-    }
-
-    private static func nonredundantDetail(
-        _ proposed: String?,
-        title: String,
-        status: String
-    ) -> String {
-        guard let proposed = compactMessage(proposed) else { return "" }
-        let occupied = Set([title, status].map(normalizedText))
-        return occupied.contains(normalizedText(proposed)) ? "" : proposed
-    }
-
-    static func displayStatus(for summaryKind: AgentOverlaySummaryKind) -> String {
-        APCLocalizedPresentation.overlayEventTitle(summaryKind)
-    }
-
-    static func summaryKind(for eventType: AgentEventKind) -> AgentOverlaySummaryKind {
-        switch eventType {
-        case .start: .start
-        case .thinking: .thinking
-        case .plan: .plan
-        case .tool: .tool
-        case .waiting: .needsInput
-        case .done: .done
-        case .failed: .failed
-        }
-    }
-
-}
-
-enum OverlayBubbleSessionPrimaryAction: Equatable {
-    case expandSessions
-    case activateSession
-
-    static func resolve(content: OverlayBubbleContent) -> Self {
-        content.isStacked ? .expandSessions : .activateSession
-    }
-}
-
-struct OverlayBubbleContent: Equatable, Identifiable {
-    var id: String
-    var source: AgentSource?
-    var agentName: String
-    var sessions: [OverlaySessionContent]
-    var isExpanded: Bool
-    var omittedSessionCount: Int
-    var isStandaloneSessionCard: Bool
-    var standaloneStackSessionCount: Int
-
-    var eventIDs: [String] { sessions.map(\.eventID) }
-    var dismissalIDs: [String] { canDismiss ? sessions.map(\.id) : [] }
-    var visibleSessions: [OverlaySessionContent] {
-        guard !sessions.isEmpty, !isOmittedSummary else { return sessions }
-        let ordered = sessions.enumerated().sorted { left, right in
-            let leftTone = OverlaySessionGroupTone(eventType: left.element.eventType)
-            let rightTone = OverlaySessionGroupTone(eventType: right.element.eventType)
-            if leftTone != rightTone {
-                return leftTone.rawValue > rightTone.rawValue
-            }
-            return left.offset < right.offset
-        }.map(\.element)
-        return isExpanded ? ordered : Array(ordered.prefix(1))
-    }
-    var sessionCount: Int { sessions.count }
-    var representedSessionCount: Int {
-        if isStandaloneSessionCard, !isExpanded {
-            return disclosureSessionCount
-        }
-        return omittedSessionCount > 0 ? omittedSessionCount : sessionCount
-    }
-    var disclosureSessionCount: Int {
-        isStandaloneSessionCard
-            ? max(sessionCount, standaloneStackSessionCount)
-            : sessionCount
-    }
-    var isOmittedSummary: Bool { omittedSessionCount > 0 }
-    var canDismiss: Bool { !isOmittedSummary }
-    var hasMultipleSessions: Bool { disclosureSessionCount > 1 }
-    var isStacked: Bool { hasMultipleSessions && !isExpanded }
-    var stackDecorationLayerCount: Int {
-        guard isStacked else { return 0 }
-        return min(
-            OverlayGeometry.bubbleCollapsedStackLayerCount,
-            max(0, disclosureSessionCount - 1)
-        )
-    }
-    var showsStackDecoration: Bool { stackDecorationLayerCount > 0 }
-    var stackDecorationDepth: CGFloat {
-        let layerOffset = isStandaloneSessionCard
-            ? OverlayGeometry.bubbleStandaloneStackLayerOffset
-            : OverlayGeometry.bubbleCollapsedStackLayerOffset
-        return CGFloat(stackDecorationLayerCount) * layerOffset
-    }
-    var statusTone: OverlaySessionGroupTone {
-        OverlaySessionGroupTone.aggregate(sessions)
-    }
-
-    /// Geometry-only fixture; idle product state intentionally emits no
-    /// bubble from AppStore.
-    static let measurementPlaceholder = OverlayBubbleContent(
-        id: "measurement-placeholder",
-        source: nil,
-        agentName: "Agent Pet Companion",
-        sessions: [.measurementPlaceholder],
-        isExpanded: true,
-        omittedSessionCount: 0,
-        isStandaloneSessionCard: false,
-        standaloneStackSessionCount: 0
-    )
-
-    static func omittedSummary(count: Int) -> OverlayBubbleContent {
-        OverlayBubbleContent(
-            id: "omitted-session-summary",
-            source: nil,
-            agentName: "Agent Pet Companion",
-            sessions: [.omittedSummary(count: count)],
-            isExpanded: true,
-            omittedSessionCount: count,
-            isStandaloneSessionCard: false,
-            standaloneStackSessionCount: 0
-        )
-    }
-
-    init(
-        id: String,
-        source: AgentSource?,
-        agentName: String,
-        sessions: [OverlaySessionContent],
-        isExpanded: Bool = true,
-        omittedSessionCount: Int = 0,
-        isStandaloneSessionCard: Bool = false,
-        standaloneStackSessionCount: Int = 0
-    ) {
-        self.id = id
-        self.source = source
-        self.agentName = agentName
-        self.sessions = sessions
-        self.isExpanded = isExpanded
-        self.omittedSessionCount = omittedSessionCount
-        self.isStandaloneSessionCard = isStandaloneSessionCard
-        self.standaloneStackSessionCount = standaloneStackSessionCount
-    }
-
-    init(source: AgentSource, states: [ActiveAgentState], isExpanded: Bool = true) {
-        id = "agent-\(source.rawValue)"
-        self.source = source
-        agentName = source.title
-        let orderedStates = Array(states
-            .enumerated()
-            .sorted(by: Self.isMoreRecentlyActivated)
-            .prefix(8))
-        sessions = orderedStates.map { entry in
-            OverlaySessionContent(state: entry.element)
-        }
-        self.isExpanded = isExpanded
-        omittedSessionCount = 0
-        isStandaloneSessionCard = false
-        standaloneStackSessionCount = 0
-    }
-
-    init(state: ActiveAgentState) {
-        self.init(source: state.source, states: [state])
-    }
-
-    init(
-        standaloneState state: ActiveAgentState,
-        stackSessionCount: Int = 1,
-        isStackExpanded: Bool = true
-    ) {
-        let session = OverlaySessionContent(state: state)
-        self.init(
-            id: "session-card-\(session.id)",
-            source: state.source,
-            agentName: state.source.title,
-            sessions: [session],
-            isExpanded: isStackExpanded,
-            isStandaloneSessionCard: true,
-            standaloneStackSessionCount: stackSessionCount
-        )
-    }
-
-    init(event: AgentEvent?) {
-        guard let event else {
-            self = .measurementPlaceholder
-            return
-        }
-        id = "agent-\(event.source.rawValue)"
-        source = event.source
-        agentName = event.source.title
-        sessions = [OverlaySessionContent(event: event)]
-        isExpanded = true
-        omittedSessionCount = 0
-        isStandaloneSessionCard = false
-        standaloneStackSessionCount = 0
-    }
-
-    private static func isMoreRecentlyActivated(
-        _ left: EnumeratedSequence<[ActiveAgentState]>.Element,
-        _ right: EnumeratedSequence<[ActiveAgentState]>.Element
-    ) -> Bool {
-        if let leftTime = left.element.sessionActivatedAt,
-           let rightTime = right.element.sessionActivatedAt,
-           leftTime != rightTime
-        {
-            return leftTime > rightTime
-        }
-        // PetCore has already projected a stable first-seen order for legacy
-        // sessions without an activation timestamp. Preserve that order and
-        // do not let a later Waiting/Failed status edge promote an old session.
-        return left.offset < right.offset
-    }
-}
-
-struct OverlayBubbleAccessibilityModel: Equatable {
-    var sessionActionLabels: [String?]
-    var sessionCloseActionLabels: [String?]
-    var closeActionLabel: String?
-    var closeActionHint: String?
-    var groupActionLabel: String?
-
-    init(content: OverlayBubbleContent, locale: String? = nil) {
-        sessionActionLabels = content.visibleSessions.map { session in
-            guard session.canOpen else { return nil }
-            guard session.source != nil else {
-                return Self.text(.overlayActionOpen, locale: locale)
-            }
-            return session.actionLabel(
-                locale: locale ?? APCLocalization.interfaceLocaleIdentifier
-            )
-        }
-        sessionCloseActionLabels = content.visibleSessions.map { _ in
-            content.canDismiss
-                ? Self.text(.overlayDismissSession, locale: locale)
-                : nil
-        }
-        closeActionLabel = content.canDismiss
-            ? Self.text(.overlayCloseBubbleAccessibility, locale: locale)
-            : nil
-        closeActionHint = content.canDismiss
-            ? Self.text(.overlayCloseBubbleHint, locale: locale)
-            : nil
-        groupActionLabel = content.hasMultipleSessions
-            ? Self.format(
-                content.isExpanded
-                    ? .overlayCollapseSessionsFormat
-                    : .overlayExpandSessionsFormat,
-                content.disclosureSessionCount,
-                locale: locale
-            )
-            : nil
-    }
-
-    private static func text(_ key: APCLocalizationKey, locale: String?) -> String {
-        locale.map { APCLocalization.text(key, locale: $0) }
-            ?? APCLocalization.text(key)
-    }
-
-    private static func format(
-        _ key: APCLocalizationKey,
-        _ count: Int,
-        locale: String?
-    ) -> String {
-        locale.map { APCLocalization.format(key, locale: $0, count) }
-            ?? APCLocalization.format(key, count)
     }
 }
