+++ crates/petcore/src/diagnostics.rs
@@ -56,7 +56,9 @@ impl DiagnosticErrorCode {
             PetCoreError::Json(_) => Self::Json,
             PetCoreError::Image(_) => Self::Image,
             PetCoreError::Zip(_) => Self::Zip,
-            PetCoreError::InvalidRequest(_) | PetCoreError::InvalidParams(_) => Self::InvalidRequest,
+            PetCoreError::InvalidRequest(_) | PetCoreError::InvalidParams(_) => {
+                Self::InvalidRequest
+            }
             PetCoreError::Validation(_) => Self::Validation,
             PetCoreError::Conflict(_) | PetCoreError::GenerationConflict { .. } => Self::Conflict,
         }
